+++ lib/dart_domainrobot_sdk.dart
@@ -2,5 +2,5 @@ library domainrobot_api;
 
 export 'src/model/models_lib.dart';
 export 'src/model/exception/exception_lib.dart';
-export 'src/DomainRobotProvider.dart';
+export 'src/DomainRobot.dart';
 export 'src/DomainRobotHeaders.dart';