+++ test/domainrobot_test.dart
@@ -3,7 +3,8 @@ import 'package:test/test.dart';
 
 void main() {
   test('Test getBasicHeaders', () {
-    var provider = DomainStudioClient('hello', 'world', '4', 'https://api.com');
+    var provider =
+        DomainStudioClient('hello', 'world', '4', 'https://api.com', '1');
     var headers = provider.mergeHeaders({'X-Domainrobot-Ctid': 'ctid-1234'});
     expect(headers.length, 4);
   });