+++ test/domainrobot_trusted_application_test.dart
@@ -8,7 +8,7 @@ import 'package:test/test.dart';
 import 'package:dart_domainrobot_sdk/dart_domainrobot_sdk.dart';
 
 void main() {
-  var provider = DomainRobotProvider('hello', 'world', '4', 'https://api.com');
+  var provider = DomainRobot('hello', 'world', '4', 'https://api.com');
 
   test('Test trustedApplicationCreate', () async {
     HttpUtils.client = MockClient((request) async {